tagso/requirements.txt
rdflib
rdflib-sqlalchemy
psycopg2-binary
flask
gunicorn
databricks-sdk
setuptools